+++ src/GRA.Controllers/ViewModel/Home/DashboardViewModel.cs
@@ -19,6 +19,9 @@ namespace GRA.Controllers.ViewModel.Home
         public string SecretCode { get; set; }
         public string SecretCodeMessage { get; set; }
 
+        public string DailyImageMessage { get; set; }
+        public string DailyImagePath { get; set; }
+
         public IEnumerable<GRA.Domain.Model.Badge> Badges { get; set; }
         public ICollection<GRA.Domain.Model.DynamicAvatarElement> DynamicAvatarElements { get; set; }
     }