+++ src/GRA.Controllers/ViewModel/Shared/DailyImageViewModel.cs
@@ -1,0 +1,8 @@
+﻿namespace GRA.Controllers.ViewModel.Shared
+{
+    public class DailyImageViewModel
+    {
+        public string DailyImageMessage { get; set; }
+        public string DailyImagePath { get; set; }
+    }
+}