+++ src/GRA.Domain.Model/Program.cs
@@ -28,5 +28,8 @@ namespace GRA.Domain.Model
 
         public int? AgeMaximum { get; set; }
         public int? AgeMinimum { get; set; }
+
+        [MaxLength(50)]
+        public string DailyImageMessage { get; set; }
     }
 }